+++ kubernetes/membership/ha_membership.go
@@ -1,25 +1,22 @@
 package membership
 
 import (
-	"time"
-
 	"github.com/Trendyol/go-dcp-client/helpers"
 	"github.com/Trendyol/go-dcp-client/membership"
 	"github.com/Trendyol/go-dcp-client/membership/info"
 )
 
 type haMembership struct {
-	info *info.Model
+	info     *info.Model
+	infoChan chan *info.Model
 }
 
 func (h *haMembership) GetInfo() *info.Model {
-	for {
-		if h.info != nil {
-			return h.info
-		}
-
-		time.Sleep(1 * time.Second)
+	if h.info != nil {
+		return h.info
 	}
+
+	return <-h.infoChan
 }
 
 func NewHaMembership(_ helpers.ConfigDCPGroupMembership, handler info.Handler) membership.Membership {
@@ -27,6 +24,9 @@ func NewHaMembership(_ helpers.ConfigDCPGroupMembership, handler info.Handler) m
 
 	handler.Subscribe(func(new *info.Model) {
 		ham.info = new
+		go func() {
+			ham.infoChan <- new
+		}()
 	})
 
 	return ham
