+++ kubernetes/membership/stateful_set_info.go
@@ -1,4 +1,4 @@
-package model
+package membership
 
 import (
 	"fmt"
